remove casting and sourceUrl

+++ datadog-monitors-downtime-handler/src/main/java/com/datadog/monitors/downtime/CreateHandler.java
@@ -75,7 +75,7 @@ public class CreateHandler extends BaseHandler<CallbackContext> {
         }
 
         // Set ID of the returned monitor, so that the read handler can request it
-        model.setId((int) (long) createdDowntime.getId());
+        model.setId(createdDowntime.getId().intValue());
         return new ReadHandler().handleRequest(proxy, request, callbackContext, logger);
     }
 }
